+++ kwin/incinerate/onAnimationBegin.js
@@ -7,8 +7,8 @@ let startPos = seed[0] > seed[1] ? [seed[0], Math.floor(seed[1] + 0.5)] :
 
 if (effect.readConfig('UsePointer', true)) {
   startPos = [
-    (effects.cursorPos.x - window.x) / window.width,
-    (effects.cursorPos.y - window.y) / window.height
+    Math.min(1.0, Math.max(0.0, (effects.cursorPos.x - window.x) / window.width)),
+    Math.min(1.0, Math.max(0.0, (effects.cursorPos.y - window.y) / window.height))
   ];
 }