+++ prefs.js
@@ -56,6 +56,8 @@ var PreferencesDialog = class PreferencesDialog {
       new Me.imports.src.Incinerate.Incinerate(),
       new Me.imports.src.Matrix.Matrix(),
       new Me.imports.src.Pixelate.Pixelate(),
+      new Me.imports.src.PixelWheel.PixelWheel(),
+      new Me.imports.src.PixelWipe.PixelWipe(),
       new Me.imports.src.SnapOfDisintegration.SnapOfDisintegration(),
       new Me.imports.src.TRexAttack.TRexAttack(),
       new Me.imports.src.TVEffect.TVEffect(),